requirements.txt
chainlit==0.6.402
langchain~=0.0.285
deepmerge==1.1.0
pyyaml~=6.0